backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
aiosqlite==0.21.0
pydantic-settings==2.11.0
httpx==0.28.1
zai-sdk
pytest==8.4.1
pytest-asyncio==1.1.0
openpyxl==3.1.5
xlrd==2.0.1
weasyprint==62.3
pydyf==0.10.0

